+++ test/p2p/clean.sh
@@ -1,5 +1,0 @@
-#! /bin/bash
-
-# clean everything
-docker rm -vf $(docker ps -aq)
-docker network rm local_testnet